Split Rolle- and ServiceProvider-Module
