+++ src/console/dbseed/db-seed-mapper.ts
@@ -2,9 +2,9 @@ import { Inject, Injectable } from '@nestjs/common';
 import { AutomapperProfile, getMapperToken } from '@automapper/nestjs';
 import { createMap, Mapper, MappingProfile } from '@automapper/core';
 import { PersonRollenZuweisungFile } from './file/person-rollen-zuweisung-file.js';
-import { PersonRollenZuweisungEntity } from '../../modules/rolle/entity/person-rollen-zuweisung.entity.js';
+import { PersonRollenZuweisungEntity } from '../../modules/service-provider/entity/person-rollen-zuweisung.entity.js';
 import { ServiceProviderZugriffFile } from './file/service-provider-zugriff-file.js';
-import { ServiceProviderZugriffEntity } from '../../modules/rolle/entity/service-provider-zugriff.entity.js';
+import { ServiceProviderZugriffEntity } from '../../modules/service-provider/entity/service-provider-zugriff.entity.js';
 import { OrganisationFile } from './file/organisation-file.js';
 import { OrganisationEntity } from '../../modules/organisation/persistence/organisation.entity.js';
 import { DataProviderFile } from './file/data-provider-file.js';
@@ -12,7 +12,7 @@ import { DataProviderEntity } from '../../persistence/data-provider.entity.js';
 import { PersonFile } from './file/person-file.js';
 import { PersonEntity } from '../../modules/person/persistence/person.entity.js';
 import { ServiceProviderFile } from './file/service-provider-file.js';
-import { ServiceProviderEntity } from '../../modules/rolle/entity/service-provider.entity.js';
+import { ServiceProviderEntity } from '../../modules/service-provider/entity/service-provider.entity.js';
 
 @Injectable()
 export class DbSeedMapper extends AutomapperProfile {
