+++ src/console/dbseed/db-seed.console.ts
@@ -10,15 +10,15 @@ import { PersonRollenZuweisungFile } from './file/person-rollen-zuweisung-file.j
 import { DbSeedService } from './db-seed.service.js';
 import { PersonFile } from './file/person-file.js';
 import { RolleEntity } from '../../modules/rolle/entity/rolle.entity.js';
-import { PersonRollenZuweisungEntity } from '../../modules/rolle/entity/person-rollen-zuweisung.entity.js';
+import { PersonRollenZuweisungEntity } from '../../modules/service-provider/entity/person-rollen-zuweisung.entity.js';
 import { ServiceProviderZugriffFile } from './file/service-provider-zugriff-file.js';
-import { ServiceProviderZugriffEntity } from '../../modules/rolle/entity/service-provider-zugriff.entity.js';
+import { ServiceProviderZugriffEntity } from '../../modules/service-provider/entity/service-provider-zugriff.entity.js';
 import { OrganisationEntity } from '../../modules/organisation/persistence/organisation.entity.js';
 import { OrganisationFile } from './file/organisation-file.js';
 import { DataProviderEntity } from '../../persistence/data-provider.entity.js';
 import { DataProviderFile } from './file/data-provider-file.js';
 import { PersonEntity } from '../../modules/person/persistence/person.entity.js';
-import { ServiceProviderEntity } from '../../modules/rolle/entity/service-provider.entity.js';
+import { ServiceProviderEntity } from '../../modules/service-provider/entity/service-provider.entity.js';
 import { ServiceProviderFile } from './file/service-provider-file.js';
 import { KeycloakUserService } from '../../modules/keycloak-administration/domain/keycloak-user.service.js';
 import { UserDo } from '../../modules/keycloak-administration/domain/user.do.js';
