+++ src/console/dbseed/file/person-rollen-zuweisung-file.ts
@@ -1,4 +1,4 @@
-import { PersonRollenZuweisungEntity } from '../../../modules/rolle/entity/person-rollen-zuweisung.entity.js';
+import { PersonRollenZuweisungEntity } from '../../../modules/service-provider/entity/person-rollen-zuweisung.entity.js';
 import { Reference } from '../db-seed.console.js';
 
 export class PersonRollenZuweisungFile extends PersonRollenZuweisungEntity {
