+++ src/console/dbseed/file/service-provider-zugriff-file.ts
@@ -1,3 +1,3 @@
-import { ServiceProviderZugriffEntity } from '../../../modules/rolle/entity/service-provider-zugriff.entity.js';
+import { ServiceProviderZugriffEntity } from '../../../modules/service-provider/entity/service-provider-zugriff.entity.js';
 
 export class ServiceProviderZugriffFile extends ServiceProviderZugriffEntity {}
