+++ src/modules/rolle/rolle.module.ts
@@ -1,34 +1,10 @@
 import { Module } from '@nestjs/common';
 import { LoggerModule } from '../../core/logging/logger.module.js';
-import { PersonRollenZuweisungRepo } from './repo/person-rollen-zuweisung.repo.js';
-import { RolleService } from './domain/rolle.service.js';
-import { PersonRollenZuweisungMapperProfile } from './mapper/person-rollen-zuweisung.mapper.profile.js';
-import { RollenBerechtigungsZuweisungRepo } from './repo/rollen-berechtigungs-zuweisung.repo.js';
-import { RolleBerechtigungsZuweisungMapperProfile } from './mapper/rolle-berechtigungs-zuweisung.mapper.profile.js';
-import { RolleRechtMapperProfile } from './mapper/rolle-recht.mapper.profile.js';
-import { ServiceProviderZugriffMapperProfile } from './mapper/service-provider-zugriff.mapper.profile.js';
-import { ServiceProviderRepo } from './repo/service-provider.repo.js';
-import { RolleRechtRepo } from './repo/rolle-recht.repo.js';
-import { ServiceProviderMapperProfile } from './mapper/service-provider.mapper.profile.js';
-import { PersonRepo } from '../person/persistence/person.repo.js';
 import { RolleRepo } from './repo/rolle.repo.js';
 
 @Module({
     imports: [LoggerModule.register(RolleModule.name)],
-    providers: [
-        PersonRepo,
-        PersonRollenZuweisungMapperProfile,
-        RolleBerechtigungsZuweisungMapperProfile,
-        RolleRechtMapperProfile,
-        ServiceProviderZugriffMapperProfile,
-        ServiceProviderMapperProfile,
-        PersonRollenZuweisungRepo,
-        RollenBerechtigungsZuweisungRepo,
-        ServiceProviderRepo,
-        RolleRechtRepo,
-        RolleService,
-        RolleRepo,
-    ],
-    exports: [RolleService, PersonRepo, RolleRepo],
+    providers: [RolleRepo],
+    exports: [RolleRepo],
 })
 export class RolleModule {}
