+++ src/modules/service-provider/mapper/rolle-berechtigungs-zuweisung.mapper.profile.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { AutomapperProfile, getMapperToken } from '@automapper/nestjs';
 import { Mapper, MappingProfile, convertUsing, createMap, forMember, mapFrom } from '@automapper/core';
-import { RolleBerechtigungsZuweisungDo } from '../domain/rolle-berechtigungs-zuweisung.do.js';
-import { RolleBerechtigungsZuweisungEntity } from '../entity/rolle-berechtigungs-zuweisung.entity.js';
+import { RolleBerechtigungsZuweisungDo } from '../../service-provider/domain/rolle-berechtigungs-zuweisung.do.js';
+import { RolleBerechtigungsZuweisungEntity } from '../../service-provider/entity/rolle-berechtigungs-zuweisung.entity.js';
 import { ServiceProviderZugriffDoRolleRechtEntityConverter } from './service-provider-zugriff-do-rolle-recht-entity.converter.js';
 
 @Injectable()
