+++ src/modules/service-provider/mapper/rolle-recht.mapper.profile.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { AutomapperProfile, getMapperToken } from '@automapper/nestjs';
 import { createMap, Mapper, MappingProfile } from '@automapper/core';
-import { RolleRechtDo } from '../domain/rolle-recht.do.js';
-import { RolleRechtEntity } from '../entity/rolle-recht.entity.js';
+import { RolleRechtDo } from '../../service-provider/domain/rolle-recht.do.js';
+import { RolleRechtEntity } from '../../service-provider/entity/rolle-recht.entity.js';
 
 @Injectable()
 export class RolleRechtMapperProfile extends AutomapperProfile {
