+++ src/modules/service-provider/mapper/service-provider-zugriff.mapper.profile.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { AutomapperProfile, getMapperToken } from '@automapper/nestjs';
 import { createMap, Mapper, MappingProfile } from '@automapper/core';
-import { ServiceProviderZugriffDo } from '../domain/service-provider-zugriff.do.js';
-import { ServiceProviderZugriffEntity } from '../entity/service-provider-zugriff.entity.js';
+import { ServiceProviderZugriffDo } from '../../service-provider/domain/service-provider-zugriff.do.js';
+import { ServiceProviderZugriffEntity } from '../../service-provider/entity/service-provider-zugriff.entity.js';
 
 @Injectable()
 export class ServiceProviderZugriffMapperProfile extends AutomapperProfile {
