+++ src/modules/service-provider/repo/rolle-recht.repo.spec.ts
@@ -11,7 +11,7 @@ import {
 } from '../../../../test/utils/index.js';
 import { RolleBerechtigungsZuweisungDo } from '../domain/rolle-berechtigungs-zuweisung.do.js';
 import { RolleRechtDo } from '../domain/rolle-recht.do.js';
-import { Rolle } from '../domain/rolle.js';
+import { Rolle } from '../../rolle/domain/rolle.js';
 import { ServiceProviderZugriffDo } from '../domain/service-provider-zugriff.do.js';
 import { ServiceProviderDo } from '../domain/service-provider.do.js';
 import { ServiceProviderZugriffEntity } from '../entity/service-provider-zugriff.entity.js';
