+++ src/persistence/entzogenes-recht.entity.ts
@@ -1,6 +1,6 @@
 import { TimestampedEntity } from './timestamped.entity.js';
 import { Entity, ManyToOne, Property } from '@mikro-orm/core';
-import { RolleRechtEntity } from '../modules/rolle/entity/rolle-recht.entity.js';
+import { RolleRechtEntity } from '../modules/service-provider/entity/rolle-recht.entity.js';
 
 @Entity({ tableName: 'entzogenes_recht' })
 export class EntzogenesRechtEntity extends TimestampedEntity {
