+++ src/persistence/system-zugriff.entity.ts
@@ -1,5 +1,5 @@
 import { Entity } from '@mikro-orm/core';
-import { RolleRechtEntity } from '../modules/rolle/entity/rolle-recht.entity.js';
+import { RolleRechtEntity } from '../modules/service-provider/entity/rolle-recht.entity.js';
 
 @Entity({ tableName: 'system_zugriff', discriminatorValue: 'systemZugriff' })
 export class SystemZugriffEntity extends RolleRechtEntity {}
